push

+++ class4/Main.java
@@ -6,7 +6,7 @@ public class Main {
         if (num == 1) return false;
         else if (num == 2) return true;
         else{
-            for (int i = 2; i <= Math.floor(Math.sqrt(num)); i++){
+            for (int i = 2; i <= (int) Math.floor(Math.sqrt(num)); i++){
                 if (num % i == 0) return false;
             }
             return true;
